test fixes and separate projects
